+++ Core/Inc/lcu/lights/brake_api.h
@@ -11,24 +11,7 @@
 #ifndef BRAKE_API_H
 #define BRAKE_API_H
 
-/*!
- * \brief           A structure that encapsulates data and functions required to
- *                  handle brake management.
- */
-struct brake_api_data {
-    bool brake_status;                    /*!< brake status false= off true= on. */
-    void (*brake_hw_update)(bool status); /*!< Function pointer to brake hardware implementation. */
-};
-
-/*!
- * \brief           Enumeration with all possible return codes of the library.
- */
-enum BrakeReturnCode {
-    BRAKE_RC_OK,               /*!< Everything is fine. */
-    BRAKE_RC_INVALID_ARGUMENT, /*!< Invalid parameter data. */
-    BRAKE_RC_NULL_POINTER     /*!< Unexpected NULL pointer. */
-};
-
+#include "brake.h"
 /*!
  * \brief           Initialize the brake_data.
  *
@@ -37,7 +20,7 @@ enum BrakeReturnCode {
  * \retval          PAL_RC_OK on success, an error otherwise:
  * \retval          PAL_RC_NULL_POINTER if any of the pointer parameters are NULL.
  */
-enum BrakeReturnCode brake_api_init(const struct brake_api_data *init_data);
+enum BrakeReturnCode brake_api_init(const struct BrakeApiHandler *init_data);
 
 /*!
  * \brief           set brake status.
