+++ Core/Src/lcu/lights/brake_api.c
@@ -11,24 +11,22 @@
 #include "brake_api.h"
 #include "eagletrt.h"
 #include <stddef.h>
-EAGLETRT_STATIC struct brake_api_data api_data;
+EAGLETRT_STATIC struct BrakeApiHandler brake_handler;
 
-//TODO: use return codes
-
-enum BrakeReturnCode brake_api_init(const struct brake_api_data *init_data) {
+enum BrakeReturnCode brake_api_init(const struct BrakeApiHandler *init_data) {
     if (init_data == NULL) {
         return BRAKE_RC_NULL_POINTER;
     }
-    api_data.brake_status = init_data->brake_status;
-    api_data.brake_hw_update = init_data->brake_hw_update;
+    brake_handler.brake_status = init_data->brake_status;
+    brake_handler.brake_hw_update = init_data->brake_hw_update;
     return BRAKE_RC_OK;
 }
 
 bool brake_api_get_status() {
-    return api_data.brake_status;
+    return brake_handler.brake_status;
 }
 
 void brake_api_set_status(bool status) {
-    api_data.brake_status = status;
-    api_data.brake_hw_update(status);
+    brake_handler.brake_status = status;
+    brake_handler.brake_hw_update(status);
 }
